+++ react/components/TabbedShelf/Tabs.js
@@ -1,7 +1,8 @@
 import React, { Component } from 'react'
 import classNames from 'classnames'
+import { withCssHandles } from 'vtex.css-handles'
 
-import tabbedShelf from './tabbedShelf.css'
+const CSS_HANDLES = ['itemContainer', 'itemContainerSelected', 'itemContainerUnselected', 'tabsContainer', 'tabsNamesContainer', 'shelfContainer', 'tabButton']
 
 class Tabs extends Component {
 
@@ -12,7 +13,7 @@ class Tabs extends Component {
   handleClick = (index) => this.setState({ selectedIndex: index })
 
   render() {
-    const { panes } = this.props
+    const { panes, cssHandles } = this.props
     const { selectedIndex } = this.state
     if (!panes || panes.length === 0) {
       return null
@@ -20,28 +21,28 @@ class Tabs extends Component {
 
     return (
       <div className="flex justify-center items-center">
-      <div className={`${tabbedShelf.tabsContainer} flex-ns pa6-ns justify-between-ns w-100-s`}>
-        <div className={`${tabbedShelf.tabsNamesContainer} flex flex-column mh6`}>
+      <div className={`${cssHandles.tabsContainer} flex-ns pa6-ns justify-between-ns w-100-s`}>
+        <div className={`${cssHandles.tabsNamesContainer} flex flex-column mh6`}>
           {panes.map((pane, index) => {
             const isSelected = index === selectedIndex
             const isLast = index === panes.length - 1
-            const itemContainer = classNames(`t-body pa4 tl bb b--muted-4 ${tabbedShelf.itemContainer}`, {
-              [`bg-base--inverted c-on-muted-1 ${tabbedShelf.itemContainerSelected}`]: isSelected,
-              [tabbedShelf.itemContainerUnselected]: !isSelected,
+            const itemContainer = classNames(`t-body pa4 tl bb b--muted-4 ${cssHandles.itemContainer}`, {
+              [`bg-base--inverted c-on-muted-1 ${cssHandles.itemContainerSelected}`]: isSelected,
+              [cssHandles.itemContainerUnselected]: !isSelected,
               'bw0': isLast,
             })
             return (
               <button 
                 type="button"
                 onClick={() => this.handleClick(index)}
-                className="bn outline-0 bg-transparent pa0"
+                className={`${cssHandles.tabButton} bn outline-0 bg-transparent pa0`}
               >
                 <div className={itemContainer} key={pane.menuItem} >{pane.menuItem}</div>
               </button>
             )
           })}
         </div>
-        <div className={`${tabbedShelf.shelfContainer} mh6-s mw-100 overflow-hidden`}>
+        <div className={`${cssHandles.shelfContainer} mh6-s mw-100 overflow-hidden`}>
           {panes[selectedIndex].render()}
         </div>
       </div>
@@ -50,4 +51,4 @@ class Tabs extends Component {
   }
 }
 
-export default Tabs
+export default withCssHandles(CSS_HANDLES)(Tabs)